refactor: relocate PhysicalTenantContext to the spring-utils module
PhysicalTenantContext is a request-scoped, servlet-layer holder (it only needs jakarta.servlet and
spring-web) — not an identity/auth concept. It was placed in the authentication module only because
that was the lowest module both the gateway layer and the auth code could already reach. Per review
feedback, move it to spring-utils: the Spring-layer commons module that is already a dependency of
the authentication and gateway modules and has no dependency on them (so no cycle).

spring-utils gains spring-web and jakarta.servlet-api (provided; supplied by the consuming web apps)
and spring-test (test). gateway-mcp now depends on camunda-spring-utils instead of
camunda-authentication for the class. All importers are repointed to
io.camunda.spring.utils.PhysicalTenantContext.
