+++ authentication/src/main/java/io/camunda/authentication/pt/PhysicalTenantPreSecurityFilter.java
@@ -7,6 +7,7 @@
  */
 package io.camunda.authentication.pt;
 
+import io.camunda.spring.utils.PhysicalTenantContext;
 import jakarta.servlet.Filter;
 import jakarta.servlet.FilterChain;
 import jakarta.servlet.ServletException;
