+++ gateways/gateway-mcp/pom.xml
@@ -35,7 +35,7 @@
     </dependency>
     <dependency>
       <groupId>io.camunda</groupId>
-      <artifactId>camunda-authentication</artifactId>
+      <artifactId>camunda-spring-utils</artifactId>
     </dependency>
     <dependency>
       <groupId>io.camunda</groupId>
