+++ gateways/gateway-mcp/src/main/java/io/camunda/gateway/mcp/tool/cluster/ClusterTools.java
@@ -7,11 +7,11 @@
  */
 package io.camunda.gateway.mcp.tool.cluster;
 
-import io.camunda.authentication.pt.PhysicalTenantContext;
 import io.camunda.gateway.mapping.http.ResponseMapper;
 import io.camunda.gateway.mcp.config.tool.CamundaMcpTool;
 import io.camunda.gateway.mcp.mapper.CallToolResultMapper;
 import io.camunda.service.registry.ServiceRegistry;
+import io.camunda.spring.utils.PhysicalTenantContext;
 import io.modelcontextprotocol.spec.McpSchema.CallToolResult;
 import org.springframework.ai.mcp.annotation.McpTool;
 import org.springframework.stereotype.Component;
