+++ zeebe/gateway-rest/src/main/java/io/camunda/zeebe/gateway/rest/annotation/PhysicalTenantId.java
@@ -7,7 +7,7 @@
  */
 package io.camunda.zeebe.gateway.rest.annotation;
 
-import io.camunda.authentication.pt.PhysicalTenantContext;
+import io.camunda.spring.utils.PhysicalTenantContext;
 import java.lang.annotation.Documented;
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
